VC: Add reset function, add Instruction dictionary and compilation, fix memory indexing GUI: Change compile button -> reset button Controller: Add functionality for compile, run/stop and reset buttons

+++ src/MainMemory.java
@@ -16,7 +16,7 @@ public class MainMemory {
         return data[index];
     }
 
-    public void write(byte index, byte value) {
+    public void write(int index, byte value) {
         data[index] = value;
     }
